Create batch_list_item component
Utilize 2 solutions to "stack" the pictures correctly: negative
margin-left and space-between. Negative margin is the default
implementation, but when there are a lot of pictures, we want to apply
space-between value to the flex-container so that the stack is
"condensed". Minimal JavaScript is required to determine when to apply the
second solution.