Add content and style to empl/settings page
The majority of the code and style is copied from empl/setup/edit since
it's similar. Maybe it can be extracted as a partial and used in both
empl/settings/show and empl/setup/edit.

Remove unneeded code from students_view.

Link employer_app_menu up to the settings page.