Extract employer settings as component
Since the form content between employer/setup/edit and
employer/settings/show are identical, we want to extract them as a
partial that can be rendered by both. The event listeners are also
extracted out.

Since we're on it, let's also extract event listeners for menu icon into
it's partial template.